fix(gql): 游标整数键经UseNumber精确还原,雪花ID分页不再失真
DecodeCursor原先json.Unmarshal到[]any,数字一律变float64;
排序键为bigint且>2^53时(雪花ID必超)游标值失真,续页
WHERE k > \$v 用错边界导致重复行或漏行,且完全静默。

改用Decoder.UseNumber解码,整数还原int64、小数还原float64,
驱动可直接编码。两处float64(2)测试夹具是把失真固化成了预期,
一并修正为int64(2)。

Constraint: 游标编码侧SQL的JSONB_BUILD_ARRAY输出精确bigint,失真只发生在Go解码侧
Rejected: 保留json.Number直接传驱动 | pgx对json.Number编码行为不可靠,显式还原更稳
Confidence: high
Scope-risk: narrow
Not-tested: 超出int64范围的无符号大整数键(实践中主键不会用)

+++ gql/compiler/pgsql/cursor_test.go
@@ -63,7 +63,7 @@ func (my *_DialectSuite) TestCursor() {
 			name:      "续页after变量游标",
 			query:     `query ($c: Cursor) { users(first: 2, after: $c, sort: { name: ASC }) { items { id } pageInfo { hasNext hasPrev } } }`,
 			variables: map[string]interface{}{"c": EncodeCursor([]any{"Bob", 2})},
-			args:      []any{EncodeCursor([]any{"Bob", 2}), "Bob", float64(2), EncodeCursor([]any{"Bob", 2})},
+			args:      []any{EncodeCursor([]any{"Bob", 2}), "Bob", int64(2), EncodeCursor([]any{"Bob", 2})}, // 整数键经UseNumber精确还原int64（雪花ID>2^53不失真）
 			expected: `SELECT JSONB_BUILD_OBJECT('users', "__sj_0"."json") AS "__root" FROM (SELECT TRUE) AS "__root_x"
 				LEFT OUTER JOIN LATERAL (
 					SELECT JSONB_BUILD_OBJECT(
